Improve e2e.Dockerfile
