+++ uui-e2e-tests/readme.md
@@ -31,7 +31,7 @@ You might need to restart computer after that.
 #### Option 1: Colima
 1. Install Colima via CLI:
     ```shell 
-    # Install "docker" because Colima needs it
+    # Install "docker" because Colima needs docker CLI client. Colima does not use docker to run containers.
     brew install docker
     
     # Install Colima itself
@@ -51,7 +51,7 @@ You might need to restart computer after that.
 
 ## Usage of alternative tools not mentioned in this guide
 By default, if "podman" is detected, then it is used to build/run containers; otherwise "docker" is used as fallback;
-To override the default behavior, you might explicitly specify any tool via ```.env.local``` file using ```UUI_DOCKER_CONTAINER_ENGINE=<cmd>``` option.
+To override the default behavior, you might explicitly specify any tool via ```.env.docker``` file using ```UUI_DOCKER_CONTAINER_ENGINE=<cmd>``` option.
 Please make sure that this tool is compatible with Docker's CLI.
 
 # Running tests in local environment
@@ -65,7 +65,7 @@ Please make sure that this tool is compatible with Docker's CLI.
      * Cons: Requires full build. Port ```5000``` might be occupied on macOS by some system utilities.
    * Any external UUI server
      * Pros: Good performance (though it depends on network speed).
-2. [Optional step] Change ```.env.local``` file to set non-standard ```UUI_APP_BASE_URL```
+2. [Optional step] Change ```.env.docker``` file to set non-standard ```UUI_APP_BASE_URL```
 
 ## NPM tasks to use
 Note: If the tasks are run for the very first time, it might take some time to download necessary docker images (up to 10 min, depends on network speed).
