+++ uui-e2e-tests/scripts/cmd/cmdRunPwDocker.ts
@@ -1,22 +1,21 @@
-import { spawnProcessSync, hasCliArg, isCmdSuccessful } from '../cliUtils';
+import { spawnProcessSync, hasCliArg } from '../cliUtils';
 import {
-    CLI_ARGS, CONTAINER_ENGINES,
+    CLI_ARGS,
     DOCKER_CONTAINER_NAME,
     DOCKER_FILES,
-    DOCKER_IMAGE_TAGS, ENV_FILES,
+    DOCKER_IMAGE_TAGS,
     YARN_TASKS,
 } from '../constants';
 import { currentMachineIpv4 } from '../ipUtils';
-import { readEnvFile } from '../envFileUtils';
-import { Logger } from '../../src/utils/logger';
+import { getContainerEngineCmd } from '../containerEngineUtils';
 
-const UUI_DOCKER_CONTAINER_MGMT = getContainerMgmtTool();
+const CONTAINER_ENGINE_CMD = getContainerEngineCmd();
 
 main();
 
 function main() {
     spawnProcessSync({
-        cmd: UUI_DOCKER_CONTAINER_MGMT,
+        cmd: CONTAINER_ENGINE_CMD,
         args: [
             'build',
             '-t',
@@ -28,7 +27,7 @@ function main() {
         exitOnErr: true,
     });
     spawnProcessSync({
-        cmd: UUI_DOCKER_CONTAINER_MGMT,
+        cmd: CONTAINER_ENGINE_CMD,
         args: [
             'rm',
             DOCKER_CONTAINER_NAME,
@@ -37,7 +36,7 @@ function main() {
     });
     const updateSnapshots = hasCliArg(CLI_ARGS.PW_DOCKER_UPDATE_SNAPSHOTS);
     spawnProcessSync({
-        cmd: UUI_DOCKER_CONTAINER_MGMT,
+        cmd: CONTAINER_ENGINE_CMD,
         args: [
             'run',
             '--name',
@@ -55,7 +54,6 @@ function main() {
             '-e',
             `UUI_DOCKER_HOST_MACHINE_IP=${currentMachineIpv4}`,
             DOCKER_IMAGE_TAGS.TEST,
-            'yarn',
             updateSnapshots ? YARN_TASKS.DOCKER_TEST_E2E_UPDATE : YARN_TASKS.DOCKER_TEST_E2E,
         ],
         exitOnErr: true,
@@ -63,38 +61,17 @@ function main() {
 }
 
 function getVolumesMapArgs() {
-    const volumesMap: Record<string, string> = {
-        './scripts': '/app/scripts',
-        './src': '/app/src',
-        './tests': '/app/tests',
-        './playwright.config.ts': '/app/playwright.config.ts',
-        './.env.ci': '/app/.env.ci',
-        './.env.local': '/app/.env.local',
-    };
-    return Object.keys(volumesMap).reduce<string[]>((acc, key) => {
-        const value = volumesMap[key];
-        acc.push('-v');
-        acc.push(`${key}:${value}`);
+    // files/folders to mount volumes
+    return [
+        './scripts',
+        './src',
+        './tests',
+        './playwright.config.ts',
+        './.env.docker',
+        './tsconfig.json',
+    ].reduce<string[]>((acc, from) => {
+        const to = `/app/${from.replace('./', '')}`;
+        acc.push('-v', `${from}:${to}`);
         return acc;
     }, []);
-}
-
-function getContainerMgmtTool(): string {
-    const envFile = readEnvFile();
-    let cmdEffective: string = envFile.UUI_DOCKER_CONTAINER_ENGINE;
-    if (cmdEffective) {
-        Logger.info(`The "${cmdEffective}" container engine is explicitly specified in "${ENV_FILES.LOCAL}"; It will be used.`);
-        return cmdEffective;
-    } else {
-        const isPodmanInstalled = isCmdSuccessful({ cmd: CONTAINER_ENGINES.podman, args: ['-v'] });
-        if (isPodmanInstalled) {
-            Logger.info(`The "${CONTAINER_ENGINES.podman}" CLI detected.`);
-            cmdEffective = CONTAINER_ENGINES.podman;
-        } else {
-            // fallback
-            cmdEffective = CONTAINER_ENGINES.docker;
-        }
-        Logger.info(`No container engine is explicitly specified in "${ENV_FILES.LOCAL}"; "${cmdEffective}" will be used.`);
-    }
-    return cmdEffective;
 }
